{{cookiecutter.github_repository_name}}/requirements.txt
Django==4.0.2
django-cors-headers==3.11.0
django-environ==0.8.1
django-health-check==3.16.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.21.2
gunicorn==20.1.0
pre-commit==2.17.0
psycopg2-binary==2.9.3
pytest-cov
pytest-django
